+++ src/main/kotlin/com/huayei/domain/exam/testPaper/repository/TestPaperRepository.kt
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 
 @Repository
 @Transactional(readOnly = true)
-interface TestPaperRepository : JpaRepository <TestPaper,Int>, JpaSpecificationExecutor <TestPaper>   {
+interface TestPaperRepository : JpaRepository <TestPaper,Long>, JpaSpecificationExecutor <TestPaper>   {
 
     fun findByCourseId(courseId : Int) : List<TestPaper>
