+++ src/main/kotlin/com/huayei/domain/exam/testQuestions/dto/TestQuestionsDto.kt
@@ -6,34 +6,34 @@ import javax.persistence.Column
 data class TestQuestionsDto (
 
     @Column(nullable = false,length = 64)
-    var questionId : Int ?= null,
+    var questionId: Long? = null,
 
     @Column(nullable = false,length = 512)
-    var questionName : String ?= null,
+    var questionName: String? = null,
 
     @Column(nullable = false)
-    var questionType : String ?= null,
+    var questionType: String? = null,
 
     @Column(nullable = false,length = 512)
-    var answer : String ?= null,
+    var answer: String? = null,
 
     @Column(length = 64)
-    var optionA : String ?= null,
+    var optionA: String? = null,
 
     @Column(length = 64)
-    var optionB : String ?= null,
+    var optionB: String? = null,
 
     @Column(length = 64)
-    var optionC : String ?= null,
+    var optionC: String? = null,
 
     @Column(length = 64)
-    var optionD : String ?= null,
+    var optionD: String? = null,
 
     @Column(nullable = false)
-    var courseId : Int ?=null
+    var courseId: Int? =null
 
 ) {
-    fun entity() : TestQuestions {
+    fun entity(): TestQuestions {
         return TestQuestions( questionName = questionName, questionType = questionType, answer = answer,
             optionA = optionA, optionB = optionB, optionC = optionC, optionD = optionD )
     }
